python/retail_agent_adk/pyproject.toml
[project]
name = "retail-agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.16.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
